Change heap_alloc parameter from uint16_t to uint32_t for bigger sizes

+++ include/kernel/heap.h
@@ -32,7 +32,7 @@ typedef struct heap_header heap_header_t;
 typedef struct heap_page heap_page_t;
 
 void init_heap(void);
-void *heap_alloc(uint16_t size);
+void *heap_alloc(uint32_t size);
 void heap_print_map(void);
 uint8_t heap_free(void *address);
